Refine Account trait with enum dispatch
