+++ RELEASES.md
@@ -15,6 +15,7 @@ None
 
 ### Internal Improvements
 - Added additional precision validations for `OrderMatchingEngine`
+- Improved `InstrumentProvider` error handling when loading (#2444), thanks @davidsblom
 - Improved order denied reason message for balance impact
 - Handle BybitErrors when updating instruments for ByBit (#2437), thanks @davidsblom
 - Retry if HttpError is raised for dYdX (#2438), thanks @davidsblom
