+++ crates/model/src/accounts/stubs.rs
@@ -16,7 +16,7 @@
 use rstest::fixture;
 
 use crate::{
-    accounts::{AccountAny, base::Account, cash::CashAccount, margin::MarginAccount},
+    accounts::{Account, AccountAny, CashAccount, MarginAccount},
     enums::{AccountType, LiquiditySide},
     events::account::{state::AccountState, stubs::*},
     identifiers::stubs::{account_id, uuid4},
