+++ crates/model/src/python/account/mod.rs
@@ -21,7 +21,7 @@ use nautilus_core::python::to_pyvalue_err;
 use pyo3::{PyObject, PyResult, Python, conversion::IntoPyObjectExt};
 
 use crate::{
-    accounts::{any::AccountAny, cash::CashAccount, margin::MarginAccount},
+    accounts::{AccountAny, CashAccount, MarginAccount},
     enums::AccountType,
 };
